requirements.txt
streamlit>=1.52.0
google-genai>=1.57.0
chromadb>=1.4.0
pysqlite3-binary